Adapt mzdb-rs code to recent thernio API changes and convert m/z from f64 to f32 - Updated for thernio API: SequencerRow -> SequenceRow, getter methods on SequenceRow, SpectrumRef access, isolation_offset (Option<f64>) - Thermo converter: use Scan.centroids() directly (newly exposed by thernio) instead of profile Spectrum for MS1 data, and thus switched to LowRes data encoding - Peak/Peakel/ExtendedPeakel/XicPeak m/z fields now use f32 - Updated detection modules, APIs (get_xic, get_nearest_peak) to use f32 - Bounding box read/filter functions use f32 for m/z parameters - Retained f64 for: DB schema fields, isolation windows, isotope calculations, R-tree queries
Databases (mzDB/peakelDB) compatibility maintained via f64->f32 conversions during reads.

+++ mzdb-rs/src/conversion/mgf.rs
@@ -167,7 +167,7 @@ impl<W: Write> MgfWriter<W> {
         }
         
         // Filter peaks by intensity threshold
-        let mut filtered_peaks: Vec<(f64, f32)> = spectrum.data.mz_array
+        let mut filtered_peaks: Vec<(f32, f32)> = spectrum.data.mz_array
             .iter()
             .zip(spectrum.data.intensity_array.iter())
             .filter(|(_mz, intensity)| **intensity >= self.options.min_intensity)
